refactor: move session envelope class
