+++ embrace-android-sdk/src/main/java/io/embrace/android/embracesdk/capture/envelope/SessionEnvelopeSource.kt
@@ -2,7 +2,7 @@ package io.embrace.android.embracesdk.capture.envelope
 
 import io.embrace.android.embracesdk.capture.envelope.session.SessionPayloadSource
 import io.embrace.android.embracesdk.internal.payload.Envelope
-import io.embrace.android.embracesdk.internal.session.SessionPayload
+import io.embrace.android.embracesdk.internal.payload.SessionPayload
 
 internal class SessionEnvelopeSource(
     private val sessionPayloadSource: SessionPayloadSource
