+++ embrace-android-sdk/src/main/java/io/embrace/android/embracesdk/capture/envelope/session/SessionPayloadSource.kt
@@ -1,6 +1,6 @@
 package io.embrace.android.embracesdk.capture.envelope.session
 
-import io.embrace.android.embracesdk.internal.session.SessionPayload
+import io.embrace.android.embracesdk.internal.payload.SessionPayload
 
 /**
  * Creates a [SessionPayload] object.
